requirements.txt
requests~=2.32.3
pytest~=8.2.2
allure-pytest~=2.13.5
selenium~=4.21.0